+++ narwhals/_arrow/expr.py
@@ -6,11 +6,17 @@ from typing import Callable
 from typing import Literal
 from typing import Sequence
 
+from narwhals._arrow.expr_cat import ArrowExprCatNamespace
+from narwhals._arrow.expr_dt import ArrowExprDateTimeNamespace
+from narwhals._arrow.expr_list import ArrowExprListNamespace
+from narwhals._arrow.expr_name import ArrowExprNameNamespace
+from narwhals._arrow.expr_str import ArrowExprStringNamespace
+from narwhals._arrow.series import ArrowSeries
 from narwhals._expression_parsing import reuse_series_implementation
-from narwhals._expression_parsing import reuse_series_namespace_implementation
 from narwhals.dependencies import get_numpy
 from narwhals.dependencies import is_numpy_array
 from narwhals.exceptions import ColumnNotFoundError
+from narwhals.typing import CompliantExpr
 from narwhals.utils import Implementation
 
 if TYPE_CHECKING:
@@ -21,9 +27,6 @@ if TYPE_CHECKING:
     from narwhals._arrow.typing import IntoArrowExpr
     from narwhals.dtypes import DType
     from narwhals.utils import Version
-
-from narwhals._arrow.series import ArrowSeries
-from narwhals.typing import CompliantExpr
 
 
 class ArrowExpr(CompliantExpr[ArrowSeries]):
@@ -560,350 +563,3 @@ class ArrowExpr(CompliantExpr[ArrowSeries]):
     @property
     def list(self: Self) -> ArrowExprListNamespace:
         return ArrowExprListNamespace(self)
-
-
-class ArrowExprCatNamespace:
-    def __init__(self: Self, expr: ArrowExpr) -> None:
-        self._compliant_expr = expr
-
-    def get_categories(self: Self) -> ArrowExpr:
-        return reuse_series_namespace_implementation(
-            self._compliant_expr, "cat", "get_categories"
-        )
-
-
-class ArrowExprDateTimeNamespace:
-    def __init__(self: Self, expr: ArrowExpr) -> None:
-        self._compliant_expr = expr
-
-    def to_string(self: Self, format: str) -> ArrowExpr:  # noqa: A002
-        return reuse_series_namespace_implementation(
-            self._compliant_expr, "dt", "to_string", format=format
-        )
-
-    def replace_time_zone(self: Self, time_zone: str | None) -> ArrowExpr:
-        return reuse_series_namespace_implementation(
-            self._compliant_expr, "dt", "replace_time_zone", time_zone=time_zone
-        )
-
-    def convert_time_zone(self: Self, time_zone: str) -> ArrowExpr:
-        return reuse_series_namespace_implementation(
-            self._compliant_expr, "dt", "convert_time_zone", time_zone=time_zone
-        )
-
-    def timestamp(self: Self, time_unit: Literal["ns", "us", "ms"] = "us") -> ArrowExpr:
-        return reuse_series_namespace_implementation(
-            self._compliant_expr, "dt", "timestamp", time_unit=time_unit
-        )
-
-    def date(self: Self) -> ArrowExpr:
-        return reuse_series_namespace_implementation(self._compliant_expr, "dt", "date")
-
-    def year(self: Self) -> ArrowExpr:
-        return reuse_series_namespace_implementation(self._compliant_expr, "dt", "year")
-
-    def month(self: Self) -> ArrowExpr:
-        return reuse_series_namespace_implementation(self._compliant_expr, "dt", "month")
-
-    def day(self: Self) -> ArrowExpr:
-        return reuse_series_namespace_implementation(self._compliant_expr, "dt", "day")
-
-    def hour(self: Self) -> ArrowExpr:
-        return reuse_series_namespace_implementation(self._compliant_expr, "dt", "hour")
-
-    def minute(self: Self) -> ArrowExpr:
-        return reuse_series_namespace_implementation(self._compliant_expr, "dt", "minute")
-
-    def second(self: Self) -> ArrowExpr:
-        return reuse_series_namespace_implementation(self._compliant_expr, "dt", "second")
-
-    def millisecond(self: Self) -> ArrowExpr:
-        return reuse_series_namespace_implementation(
-            self._compliant_expr, "dt", "millisecond"
-        )
-
-    def microsecond(self: Self) -> ArrowExpr:
-        return reuse_series_namespace_implementation(
-            self._compliant_expr, "dt", "microsecond"
-        )
-
-    def nanosecond(self: Self) -> ArrowExpr:
-        return reuse_series_namespace_implementation(
-            self._compliant_expr, "dt", "nanosecond"
-        )
-
-    def ordinal_day(self: Self) -> ArrowExpr:
-        return reuse_series_namespace_implementation(
-            self._compliant_expr, "dt", "ordinal_day"
-        )
-
-    def weekday(self: Self) -> ArrowExpr:
-        return reuse_series_namespace_implementation(
-            self._compliant_expr, "dt", "weekday"
-        )
-
-    def total_minutes(self: Self) -> ArrowExpr:
-        return reuse_series_namespace_implementation(
-            self._compliant_expr, "dt", "total_minutes"
-        )
-
-    def total_seconds(self: Self) -> ArrowExpr:
-        return reuse_series_namespace_implementation(
-            self._compliant_expr, "dt", "total_seconds"
-        )
-
-    def total_milliseconds(self: Self) -> ArrowExpr:
-        return reuse_series_namespace_implementation(
-            self._compliant_expr, "dt", "total_milliseconds"
-        )
-
-    def total_microseconds(self: Self) -> ArrowExpr:
-        return reuse_series_namespace_implementation(
-            self._compliant_expr, "dt", "total_microseconds"
-        )
-
-    def total_nanoseconds(self: Self) -> ArrowExpr:
-        return reuse_series_namespace_implementation(
-            self._compliant_expr, "dt", "total_nanoseconds"
-        )
-
-
-class ArrowExprStringNamespace:
-    def __init__(self: Self, expr: ArrowExpr) -> None:
-        self._compliant_expr = expr
-
-    def len_chars(self: Self) -> ArrowExpr:
-        return reuse_series_namespace_implementation(
-            self._compliant_expr, "str", "len_chars"
-        )
-
-    def replace(
-        self: Self, pattern: str, value: str, *, literal: bool, n: int
-    ) -> ArrowExpr:
-        return reuse_series_namespace_implementation(
-            self._compliant_expr,
-            "str",
-            "replace",
-            pattern=pattern,
-            value=value,
-            literal=literal,
-            n=n,
-        )
-
-    def replace_all(self: Self, pattern: str, value: str, *, literal: bool) -> ArrowExpr:
-        return reuse_series_namespace_implementation(
-            self._compliant_expr,
-            "str",
-            "replace_all",
-            pattern=pattern,
-            value=value,
-            literal=literal,
-        )
-
-    def strip_chars(self: Self, characters: str | None) -> ArrowExpr:
-        return reuse_series_namespace_implementation(
-            self._compliant_expr, "str", "strip_chars", characters=characters
-        )
-
-    def starts_with(self: Self, prefix: str) -> ArrowExpr:
-        return reuse_series_namespace_implementation(
-            self._compliant_expr, "str", "starts_with", prefix=prefix
-        )
-
-    def ends_with(self: Self, suffix: str) -> ArrowExpr:
-        return reuse_series_namespace_implementation(
-            self._compliant_expr, "str", "ends_with", suffix=suffix
-        )
-
-    def contains(self, pattern: str, *, literal: bool) -> ArrowExpr:
-        return reuse_series_namespace_implementation(
-            self._compliant_expr, "str", "contains", pattern=pattern, literal=literal
-        )
-
-    def slice(self: Self, offset: int, length: int | None) -> ArrowExpr:
-        return reuse_series_namespace_implementation(
-            self._compliant_expr, "str", "slice", offset=offset, length=length
-        )
-
-    def to_datetime(self: Self, format: str | None) -> ArrowExpr:  # noqa: A002
-        return reuse_series_namespace_implementation(
-            self._compliant_expr, "str", "to_datetime", format=format
-        )
-
-    def to_uppercase(self: Self) -> ArrowExpr:
-        return reuse_series_namespace_implementation(
-            self._compliant_expr, "str", "to_uppercase"
-        )
-
-    def to_lowercase(self: Self) -> ArrowExpr:
-        return reuse_series_namespace_implementation(
-            self._compliant_expr, "str", "to_lowercase"
-        )
-
-
-class ArrowExprNameNamespace:
-    def __init__(self: Self, expr: ArrowExpr) -> None:
-        self._compliant_expr = expr
-
-    def keep(self: Self) -> ArrowExpr:
-        root_names = self._compliant_expr._root_names
-
-        if root_names is None:
-            msg = (
-                "Anonymous expressions are not supported in `.name.keep`.\n"
-                "Instead of `nw.all()`, try using a named expression, such as "
-                "`nw.col('a', 'b')`\n"
-            )
-            raise ValueError(msg)
-
-        return self._compliant_expr.__class__(
-            lambda df: [
-                series.alias(name)
-                for series, name in zip(self._compliant_expr._call(df), root_names)
-            ],
-            depth=self._compliant_expr._depth,
-            function_name=self._compliant_expr._function_name,
-            root_names=root_names,
-            output_names=root_names,
-            backend_version=self._compliant_expr._backend_version,
-            version=self._compliant_expr._version,
-            kwargs=self._compliant_expr._kwargs,
-        )
-
-    def map(self: Self, function: Callable[[str], str]) -> ArrowExpr:
-        root_names = self._compliant_expr._root_names
-
-        if root_names is None:
-            msg = (
-                "Anonymous expressions are not supported in `.name.map`.\n"
-                "Instead of `nw.all()`, try using a named expression, such as "
-                "`nw.col('a', 'b')`\n"
-            )
-            raise ValueError(msg)
-
-        output_names = [function(str(name)) for name in root_names]
-
-        return self._compliant_expr.__class__(
-            lambda df: [
-                series.alias(name)
-                for series, name in zip(self._compliant_expr._call(df), output_names)
-            ],
-            depth=self._compliant_expr._depth,
-            function_name=self._compliant_expr._function_name,
-            root_names=root_names,
-            output_names=output_names,
-            backend_version=self._compliant_expr._backend_version,
-            version=self._compliant_expr._version,
-            kwargs={**self._compliant_expr._kwargs, "function": function},
-        )
-
-    def prefix(self: Self, prefix: str) -> ArrowExpr:
-        root_names = self._compliant_expr._root_names
-        if root_names is None:
-            msg = (
-                "Anonymous expressions are not supported in `.name.prefix`.\n"
-                "Instead of `nw.all()`, try using a named expression, such as "
-                "`nw.col('a', 'b')`\n"
-            )
-            raise ValueError(msg)
-
-        output_names = [prefix + str(name) for name in root_names]
-        return self._compliant_expr.__class__(
-            lambda df: [
-                series.alias(name)
-                for series, name in zip(self._compliant_expr._call(df), output_names)
-            ],
-            depth=self._compliant_expr._depth,
-            function_name=self._compliant_expr._function_name,
-            root_names=root_names,
-            output_names=output_names,
-            backend_version=self._compliant_expr._backend_version,
-            version=self._compliant_expr._version,
-            kwargs={**self._compliant_expr._kwargs, "prefix": prefix},
-        )
-
-    def suffix(self: Self, suffix: str) -> ArrowExpr:
-        root_names = self._compliant_expr._root_names
-        if root_names is None:
-            msg = (
-                "Anonymous expressions are not supported in `.name.suffix`.\n"
-                "Instead of `nw.all()`, try using a named expression, such as "
-                "`nw.col('a', 'b')`\n"
-            )
-            raise ValueError(msg)
-
-        output_names = [str(name) + suffix for name in root_names]
-
-        return self._compliant_expr.__class__(
-            lambda df: [
-                series.alias(name)
-                for series, name in zip(self._compliant_expr._call(df), output_names)
-            ],
-            depth=self._compliant_expr._depth,
-            function_name=self._compliant_expr._function_name,
-            root_names=root_names,
-            output_names=output_names,
-            backend_version=self._compliant_expr._backend_version,
-            version=self._compliant_expr._version,
-            kwargs={**self._compliant_expr._kwargs, "suffix": suffix},
-        )
-
-    def to_lowercase(self: Self) -> ArrowExpr:
-        root_names = self._compliant_expr._root_names
-
-        if root_names is None:
-            msg = (
-                "Anonymous expressions are not supported in `.name.to_lowercase`.\n"
-                "Instead of `nw.all()`, try using a named expression, such as "
-                "`nw.col('a', 'b')`\n"
-            )
-            raise ValueError(msg)
-        output_names = [str(name).lower() for name in root_names]
-
-        return self._compliant_expr.__class__(
-            lambda df: [
-                series.alias(name)
-                for series, name in zip(self._compliant_expr._call(df), output_names)
-            ],
-            depth=self._compliant_expr._depth,
-            function_name=self._compliant_expr._function_name,
-            root_names=root_names,
-            output_names=output_names,
-            backend_version=self._compliant_expr._backend_version,
-            version=self._compliant_expr._version,
-            kwargs=self._compliant_expr._kwargs,
-        )
-
-    def to_uppercase(self: Self) -> ArrowExpr:
-        root_names = self._compliant_expr._root_names
-
-        if root_names is None:
-            msg = (
-                "Anonymous expressions are not supported in `.name.to_uppercase`.\n"
-                "Instead of `nw.all()`, try using a named expression, such as "
-                "`nw.col('a', 'b')`\n"
-            )
-            raise ValueError(msg)
-        output_names = [str(name).upper() for name in root_names]
-
-        return self._compliant_expr.__class__(
-            lambda df: [
-                series.alias(name)
-                for series, name in zip(self._compliant_expr._call(df), output_names)
-            ],
-            depth=self._compliant_expr._depth,
-            function_name=self._compliant_expr._function_name,
-            root_names=root_names,
-            output_names=output_names,
-            backend_version=self._compliant_expr._backend_version,
-            version=self._compliant_expr._version,
-            kwargs=self._compliant_expr._kwargs,
-        )
-
-
-class ArrowExprListNamespace:
-    def __init__(self: Self, expr: ArrowExpr) -> None:
-        self._expr = expr
-
-    def len(self: Self) -> ArrowExpr:
-        return reuse_series_namespace_implementation(self._expr, "list", "len")
